+++ runtime/src/abis_grp_runtime/e2e/__init__.py
@@ -1,8 +1,7 @@
 """Reference runtime end-to-end integration service."""
 
-from abis_grp_runtime.e2e.service import GrokE2EResponse, GrokE2EService
+from abis_grp_runtime.e2e.service import GrokE2EService
 
 __all__ = [
-    "GrokE2EResponse",
     "GrokE2EService",
 ]